Update index.md
- Corrected note about Quest URL resolver, VRChat has one of Quest now as well, added in a recent update
- Corrected links to prefabs (they have moved to the path of the package now). 

IMPORTANT NOTE 1: VRChat forgot to add the AVPro prefab to the menu and to the Udon example scene, otherwise it would be better to point users there.

IMPORTANT NOTE 2: Those prefabs break when adding them to a scene and must be manually corrected by the users, so asking the user to use udon prefabs is bad (prefabs and Udon doesn't go well together, they always lose their references). It would be better to have them in the example scene instead, unpacked, and ask the user to go there. Providing Udon Prefabs is bad practice. Tested by 2 independent users with the newest SDK.